pyproject.toml
[tool.poetry]
name = "pytrader"
version = "0.1.0"
description = "CLI for identifying a creating trades"
authors = ["AJ <11035621+axlj45@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
joblib = "^1.4.2"
pandas = "^2.2.2"
pandas-datareader = "^0.10.0"
scikit-learn = "^1.5.0"
pandas-ta = "^0.3.14b0"
statsmodels = "^0.14.2"
yfinance = "0.2.31"
plotly = "^5.22.0"
matplotlib = "^3.9.0"
click = "^8.1.7"
pytz = "^2024.1"
alpaca-py = "^0.21.2"
python-dotenv = "^1.0.1"
firebase-admin = "^6.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
